+++ Primitives/interface/FormatString.h
@@ -57,18 +57,18 @@ namespace Diligent
     template<typename Type>
     struct MemorySizeFormatter
     {
-        MemorySizeFormatter(Type _size, size_t _precision, Type _ref_size) : 
+        MemorySizeFormatter(Type _size, int _precision, Type _ref_size) : 
             size     (_size),
             precision(_precision),
             ref_size (_ref_size)
         {}
-        Type   size      = 0;
-        size_t precision = 0;
-        Type   ref_size  = 0;
+        Type size      = 0;
+        int  precision = 0;
+        Type ref_size  = 0;
     };
 
     template<typename Type>
-    MemorySizeFormatter<Type> FormatMemorySize(Type _size, size_t _precision = 0, Type _ref_size = 0)
+    MemorySizeFormatter<Type> FormatMemorySize(Type _size, int _precision = 0, Type _ref_size = 0)
     {
         return MemorySizeFormatter<Type>{_size, _precision, _ref_size};
     }
